рефакторинг

+++ src/controller/tour.php
@@ -2,6 +2,11 @@
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\JsonResponse;
 
+/**
+ * @var Request $request
+ */
+$request = new Request();
+
 $app
 	->get('/афиша', function () use ($app, $service) {
 		$tours = $service->selectAll('tours');
@@ -13,18 +18,14 @@ $app
 	->bind('tour');
 
 $app
-	->post('/add_tour', function (Request $request) use ($app) {
-		$status[] = '';
-		$date = trim($request->get('date'));
-		$city = trim($request->get('city'));
-		$place = trim($request->get('place'));
+	->post('/add_tour', function () use ($app, $request) {
 
 		try {
 			$app['db']->insert(
 				'tours', [
-				'date' => $date,
-				'city' => $city,
-				'location' => $place,
+				'date' => trim($request->request->get('date')),
+				'city' => trim($request->request->get('city')),
+				'location' => trim($request->request->get('place')),
 			]);
 			$status = [
 				'type' => 'success',
@@ -43,34 +44,33 @@ $app
 	->bind('add_tour');
 
 $app
-	->post('/delete_tour', function () use ($app) {
-		$id = $_POST['tour_id'];
+	->post('/delete_tour', function () use ($app, $request) {
+        if ($request->isMethod('POST')) {
+            $id = $request->request->get('tour_id');
+            try {
+                $app['db']->delete('tours', ['id' => $id]);
 
-		try {
-			$app['db']->delete('tours', ['id' => $id]);
-			return $app->redirect($app["url_generator"]->generate("dashboard"));
-		} catch (\Exception $e) {
-			throw new Exception($e->getMessage());
-		}
+                return $app->redirect(
+                    $app["url_generator"]->generate("dashboard")
+                );
+            } catch (\Exception $e) {
+                throw new Exception($e->getMessage());
+            }
+        }
 	})
 	->bind('delete_tour');
 
 $app
-	->post('/edit_tour', function (Request $request) use ($app, $service) {
-		$status[] = '';
-		$id = trim($request->get('id'));
-		$date = trim($request->get('date'));
-		$city = trim($request->get('city'));
-		$place = trim($request->get('place'));
+	->post('/edit_tour', function () use ($app, $service, $request) {
 
 		try {
 			$app['db']->update(
 				'tours', [
-				'date' => $date,
-				'city' => $city,
-				'location' => $place,
+				'date' => trim($request->request->get('date')),
+				'city' => trim($request->request->get('city')),
+				'location' => trim($request->request->get('place')),
 			], [
-				'id' => $id,
+				'id' => trim($request->request->get('id')),
 			]);
 			$status = [
 				'type' => 'success',
